move Decoder annotation to specifier subpackage
