--- cloud-minecraft-extras/src/main/java/org/incendo/cloud/minecraft/extras/annotation/package-info.java
@@ -1,4 +1,0 @@
-/**
- * Annotations for cloud-minecraft-extras.
- */
-package org.incendo.cloud.minecraft.extras.annotation;
