+++ cloud-minecraft-extras/src/main/java/org/incendo/cloud/minecraft/extras/annotation/specifier/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Annotations for cloud-minecraft-extras.
+ */
+package org.incendo.cloud.minecraft.extras.annotation.specifier;
